bump code and dependencies to lexical 0.16.0
