+++ docs/package.json
@@ -7,34 +7,34 @@
   },
   "type": "module",
   "dependencies": {
-    "@lexical/clipboard": "0.14.3",
-    "@lexical/code": "0.14.3",
-    "@lexical/dragon": "0.14.3",
-    "@lexical/hashtag": "0.14.3",
-    "@lexical/history": "0.14.3",
-    "@lexical/link": "0.14.3",
-    "@lexical/list": "0.14.3",
-    "@lexical/mark": "0.14.3",
-    "@lexical/markdown": "0.14.3",
-    "@lexical/overflow": "0.14.3",
-    "@lexical/plain-text": "0.14.3",
-    "@lexical/rich-text": "0.14.3",
-    "@lexical/selection": "0.14.3",
-    "@lexical/table": "0.14.3",
-    "@lexical/text": "0.14.3",
-    "@lexical/utils": "0.14.3",
-    "@lexical/yjs": "0.14.3",
-    "lexical": "0.14.3",
+    "@lexical/clipboard": "0.16.0",
+    "@lexical/code": "0.16.0",
+    "@lexical/dragon": "0.16.0",
+    "@lexical/hashtag": "0.16.0",
+    "@lexical/history": "0.16.0",
+    "@lexical/link": "0.16.0",
+    "@lexical/list": "0.16.0",
+    "@lexical/mark": "0.16.0",
+    "@lexical/markdown": "0.16.0",
+    "@lexical/overflow": "0.16.0",
+    "@lexical/plain-text": "0.16.0",
+    "@lexical/rich-text": "0.16.0",
+    "@lexical/selection": "0.16.0",
+    "@lexical/table": "0.16.0",
+    "@lexical/text": "0.16.0",
+    "@lexical/utils": "0.16.0",
+    "@lexical/yjs": "0.16.0",
+    "lexical": "0.16.0",
     "lexical-solid": "workspace:*"
   },
   "devDependencies": {
-    "@solidjs/meta": "^0.29.3",
-    "@solidjs/router": "^0.13.1",
-    "@solidjs/start": "1.0.0-rc.0",
-    "solid-js": "^1.8.16",
-    "typescript": "^5.4.3",
-    "vinxi": "^0.3.11",
-    "vite": "^5.2.8"
+    "@solidjs/meta": "^0.29.4",
+    "@solidjs/router": "^0.13.6",
+    "@solidjs/start": "^1.0.2",
+    "solid-js": "^1.8.17",
+    "typescript": "^5.5.2",
+    "vinxi": "^0.3.12",
+    "vite": "^5.3.1"
   },
   "engines": {
     "node": ">=20"
