+++ lexical-solid/package.json
@@ -39,42 +39,42 @@
     "build": "node build.js"
   },
   "dependencies": {
-    "@lexical/clipboard": "~0.14.3",
-    "@lexical/code": "~0.14.3",
-    "@lexical/dragon": "~0.14.3",
-    "@lexical/hashtag": "~0.14.3",
-    "@lexical/history": "~0.14.3",
-    "@lexical/html": "~0.14.3",
-    "@lexical/link": "~0.14.3",
-    "@lexical/list": "~0.14.3",
-    "@lexical/mark": "~0.14.3",
-    "@lexical/markdown": "~0.14.3",
-    "@lexical/overflow": "~0.14.3",
-    "@lexical/plain-text": "~0.14.3",
-    "@lexical/rich-text": "~0.14.3",
-    "@lexical/selection": "~0.14.3",
-    "@lexical/table": "~0.14.3",
-    "@lexical/text": "~0.14.3",
-    "@lexical/utils": "~0.14.3",
-    "@lexical/yjs": "~0.14.3",
-    "lexical": "~0.14.3"
+    "@lexical/clipboard": "~0.16.0",
+    "@lexical/code": "~0.16.0",
+    "@lexical/dragon": "~0.16.0",
+    "@lexical/hashtag": "~0.16.0",
+    "@lexical/history": "~0.16.0",
+    "@lexical/html": "~0.16.0",
+    "@lexical/link": "~0.16.0",
+    "@lexical/list": "~0.16.0",
+    "@lexical/mark": "~0.16.0",
+    "@lexical/markdown": "~0.16.0",
+    "@lexical/overflow": "~0.16.0",
+    "@lexical/plain-text": "~0.16.0",
+    "@lexical/rich-text": "~0.16.0",
+    "@lexical/selection": "~0.16.0",
+    "@lexical/table": "~0.16.0",
+    "@lexical/text": "~0.16.0",
+    "@lexical/utils": "~0.16.0",
+    "@lexical/yjs": "~0.16.0",
+    "lexical": "~0.16.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.9",
-    "@babel/preset-typescript": "^7.23.3",
+    "@babel/core": "^7.24.7",
+    "@babel/preset-typescript": "^7.24.7",
     "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-commonjs": "^25.0.8",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "babel-preset-solid": "^1.8.12",
+    "babel-preset-solid": "^1.8.17",
     "cross-env": "^7.0.3",
     "fast-glob": "^3.3.2",
     "npm-run-all": "^4.1.5",
-    "rimraf": "^5.0.5",
-    "rollup": "^4.13.0",
-    "solid-js": "^1.8.15",
-    "typescript": "^5.4.2",
+    "rimraf": "^5.0.7",
+    "rollup": "^4.18.0",
+    "solid-js": "^1.8.17",
+    "typescript": "^5.5.2",
     "utility-types": "3.11.0",
-    "yjs": "^13.6.14"
+    "yjs": "^13.6.18"
   },
   "peerDependencies": {
     "solid-js": "^1.7.11",
