+++ lexical-solid/src/LexicalAutoLinkPlugin.tsx
@@ -171,7 +171,7 @@ function extractMatchingNodes(
   ];
 }
 
-function createAutoLinkNode(
+function $createAutoLinkNode_(
   nodes: TextNode[],
   startIndex: number,
   endIndex: number,
@@ -192,6 +192,7 @@ function createAutoLinkNode(
     const textNode = $createTextNode(match.text);
     textNode.setFormat(linkTextNode.getFormat());
     textNode.setDetail(linkTextNode.getDetail());
+    textNode.setStyle(linkTextNode.getStyle());
     linkNode.append(textNode);
     linkTextNode.replace(linkNode);
     return remainingTextNode;
@@ -232,6 +233,7 @@ function createAutoLinkNode(
     const textNode = $createTextNode(firstLinkTextNode.getTextContent());
     textNode.setFormat(firstLinkTextNode.getFormat());
     textNode.setDetail(firstLinkTextNode.getDetail());
+    textNode.setStyle(firstLinkTextNode.getStyle());
     linkNode.append(textNode, ...linkNodes);
     // it does not preserve caret position if caret was at the first text node
     // so we need to restore caret position
@@ -248,7 +250,7 @@ function createAutoLinkNode(
   return undefined;
 }
 
-function handleLinkCreation(
+function $handleLinkCreation(
   nodes: TextNode[],
   matchers: Array<LinkMatcher>,
   onChange: ChangeHandler
@@ -282,7 +284,7 @@ function handleLinkCreation(
 
       const actualMatchStart = invalidMatchEnd + matchStart - matchingOffset;
       const actualMatchEnd = invalidMatchEnd + matchEnd - matchingOffset;
-      const remainingTextNode = createAutoLinkNode(
+      const remainingTextNode = $createAutoLinkNode_(
         matchingNodes,
         actualMatchStart,
         actualMatchEnd,
@@ -443,7 +445,7 @@ function useAutoLink(
                 !$isAutoLinkNode(previous))
             ) {
               const textNodesToMatch = getTextNodesToMatch(textNode);
-              handleLinkCreation(textNodesToMatch, matchers, onChangeWrapped);
+              $handleLinkCreation(textNodesToMatch, matchers, onChangeWrapped);
             }
 
             handleBadNeighbors(textNode, matchers, onChangeWrapped);
