+++ lexical-solid/src/LexicalCheckListPlugin.tsx
@@ -218,6 +218,7 @@ function handleClick(event: Event) {
       });
     }
   });
+  event.stopPropagation(); // Prevents potential parent-editor listeners from firing, causing the checkbox toggle to be undone
 }
 
 function handlePointerDown(event: PointerEvent) {
@@ -225,6 +226,7 @@ function handlePointerDown(event: PointerEvent) {
     // Prevents caret moving when clicking on check mark
     event.preventDefault();
   });
+  event.stopPropagation();
 }
 
 function findEditor(target: Node) {
