+++ lexical-solid/src/LexicalClearEditorPlugin.tsx
@@ -3,6 +3,7 @@ import {
   $createParagraphNode,
   $getRoot,
   $getSelection,
+  $isRangeSelection,
   CLEAR_EDITOR_COMMAND,
   COMMAND_PRIORITY_EDITOR,
 } from "lexical";
@@ -27,6 +28,9 @@ export function ClearEditorPlugin(props: Props): JSX.Element {
             root.append(paragraph);
             if (selection !== null) {
               paragraph.select();
+            }
+            if ($isRangeSelection(selection)) {
+              selection.format = 0;
             }
           } else {
             props.onClear();
