+++ lexical-solid/src/LexicalTableOfContents.tsx
@@ -1,215 +1,12 @@
-import type { LexicalEditor, NodeKey, NodeMutation } from "lexical";
-
-import { useLexicalComposerContext } from "./LexicalComposerContext";
+/** @deprecated moved to lexical-solid/LexicalTableOfContentsPlugin */
 import {
-  $isHeadingNode,
-  HeadingNode,
-  HeadingTagType,
-} from "@lexical/rich-text";
-import { $getNodeByKey, $getRoot, TextNode } from "lexical";
-import {
-  type JSX,
-  createSignal,
-  Accessor,
-  createEffect,
-  onCleanup,
-} from "solid-js";
+  type TableOfContentsEntry,
+  TableOfContentsPlugin,
+} from "./LexicalTableOfContentsPlugin";
 
-export type TableOfContentsEntry = [
-  key: NodeKey,
-  text: string,
-  tag: HeadingTagType
-];
+/** @deprecated use the module lexical-solid/LexicalTableOfContentsPlugin */
+export { type TableOfContentsEntry };
 
-function toEntry(heading: HeadingNode): TableOfContentsEntry {
-  return [heading.getKey(), heading.getTextContent(), heading.getTag()];
-}
-
-function $insertHeadingIntoTableOfContents(
-  prevHeading: HeadingNode | null,
-  newHeading: HeadingNode | null,
-  currentTableOfContents: Array<TableOfContentsEntry>
-): Array<TableOfContentsEntry> {
-  if (newHeading === null) {
-    return currentTableOfContents;
-  }
-  const newEntry: TableOfContentsEntry = toEntry(newHeading);
-  let newTableOfContents: Array<TableOfContentsEntry> = [];
-  if (prevHeading === null) {
-    newTableOfContents = [newEntry, ...currentTableOfContents];
-  } else {
-    for (let i = 0; i < currentTableOfContents.length; i++) {
-      const key = currentTableOfContents[i][0];
-      newTableOfContents.push(currentTableOfContents[i]);
-      if (key === prevHeading.getKey() && key !== newHeading.getKey()) {
-        newTableOfContents.push(newEntry);
-      }
-    }
-  }
-  return newTableOfContents;
-}
-
-function $deleteHeadingFromTableOfContents(
-  key: NodeKey,
-  currentTableOfContents: Array<TableOfContentsEntry>
-): Array<TableOfContentsEntry> {
-  const newTableOfContents = [];
-  for (const heading of currentTableOfContents) {
-    if (heading[0] !== key) {
-      newTableOfContents.push(heading);
-    }
-  }
-  return newTableOfContents;
-}
-
-function $updateHeadingInTableOfContents(
-  heading: HeadingNode,
-  currentTableOfContents: Array<TableOfContentsEntry>
-): Array<TableOfContentsEntry> {
-  const newTableOfContents: Array<TableOfContentsEntry> = [];
-  for (const oldHeading of currentTableOfContents) {
-    if (oldHeading[0] === heading.getKey()) {
-      newTableOfContents.push(toEntry(heading));
-    } else {
-      newTableOfContents.push(oldHeading);
-    }
-  }
-  return newTableOfContents;
-}
-
-/**
- * Returns the updated table of contents, placing the given `heading` before the given `prevHeading`. If `prevHeading`
- * is undefined, `heading` is placed at the start of table of contents
- */
-function $updateHeadingPosition(
-  prevHeading: HeadingNode | null,
-  heading: HeadingNode,
-  currentTableOfContents: Array<TableOfContentsEntry>
-): Array<TableOfContentsEntry> {
-  const newTableOfContents: Array<TableOfContentsEntry> = [];
-  const newEntry: TableOfContentsEntry = toEntry(heading);
-
-  if (!prevHeading) {
-    newTableOfContents.push(newEntry);
-  }
-  for (const oldHeading of currentTableOfContents) {
-    if (oldHeading[0] === heading.getKey()) {
-      continue;
-    }
-    newTableOfContents.push(oldHeading);
-    if (prevHeading && oldHeading[0] === prevHeading.getKey()) {
-      newTableOfContents.push(newEntry);
-    }
-  }
-
-  return newTableOfContents;
-}
-
-type Props = {
-  children: (
-    values: Accessor<Array<TableOfContentsEntry>>,
-    editor: LexicalEditor
-  ) => JSX.Element;
-};
-
-export function LexicalTableOfContentsPlugin({
-  children,
-}: Props): JSX.Element {
-  const [tableOfContents, setTableOfContents] = createSignal<
-    Array<TableOfContentsEntry>
-  >([]);
-  const [editor] = useLexicalComposerContext();
-  createEffect(() => {
-    // Set table of contents initial state
-    let currentTableOfContents: Array<TableOfContentsEntry> = [];
-    editor.getEditorState().read(() => {
-      const root = $getRoot();
-      const rootChildren = root.getChildren();
-      for (const child of rootChildren) {
-        if ($isHeadingNode(child)) {
-          currentTableOfContents.push([
-            child.getKey(),
-            child.getTextContent(),
-            child.getTag(),
-          ]);
-        }
-      }
-      setTableOfContents(currentTableOfContents);
-    });
-
-    // Listen to updates to heading mutations and update state
-    const removeHeaderMutationListener = editor.registerMutationListener(
-      HeadingNode,
-      (mutatedNodes: Map<string, NodeMutation>) => {
-        editor.getEditorState().read(() => {
-          for (const [nodeKey, mutation] of mutatedNodes) {
-            if (mutation === "created") {
-              const newHeading = $getNodeByKey<HeadingNode>(nodeKey);
-              if (newHeading !== null) {
-                let prevHeading = newHeading.getPreviousSibling();
-                while (prevHeading !== null && !$isHeadingNode(prevHeading)) {
-                  prevHeading = prevHeading.getPreviousSibling();
-                }
-                currentTableOfContents = $insertHeadingIntoTableOfContents(
-                  prevHeading,
-                  newHeading,
-                  currentTableOfContents
-                );
-              }
-            } else if (mutation === "destroyed") {
-              currentTableOfContents = $deleteHeadingFromTableOfContents(
-                nodeKey,
-                currentTableOfContents
-              );
-            } else if (mutation === "updated") {
-              const newHeading = $getNodeByKey<HeadingNode>(nodeKey);
-              if (newHeading !== null) {
-                let prevHeading = newHeading.getPreviousSibling();
-                while (prevHeading !== null && !$isHeadingNode(prevHeading)) {
-                  prevHeading = prevHeading.getPreviousSibling();
-                }
-                currentTableOfContents = $updateHeadingPosition(
-                  prevHeading,
-                  newHeading,
-                  currentTableOfContents
-                );
-              }
-            }
-          }
-          setTableOfContents(currentTableOfContents);
-        });
-      }
-    );
-
-    // Listen to text node mutation updates
-    const removeTextNodeMutationListener = editor.registerMutationListener(
-      TextNode,
-      (mutatedNodes: Map<string, NodeMutation>) => {
-        editor.getEditorState().read(() => {
-          for (const [nodeKey, mutation] of mutatedNodes) {
-            if (mutation === "updated") {
-              const currNode = $getNodeByKey(nodeKey);
-              if (currNode !== null) {
-                const parentNode = currNode.getParentOrThrow();
-                if ($isHeadingNode(parentNode)) {
-                  currentTableOfContents = $updateHeadingInTableOfContents(
-                    parentNode,
-                    currentTableOfContents
-                  );
-                  setTableOfContents(currentTableOfContents);
-                }
-              }
-            }
-          }
-        });
-      }
-    );
-
-    onCleanup(() => {
-      removeHeaderMutationListener();
-      removeTextNodeMutationListener();
-    });
-  });
-
-  return children(tableOfContents, editor);
-}
+/** @deprecated use module lexical-solid/LexicalTableOfContentsPlugin */
+// eslint-disable-next-line no-restricted-exports
+export default TableOfContentsPlugin;
