+++ lexical-solid/src/shared/LexicalMenu.tsx
@@ -480,7 +480,9 @@ export function useMenuAnchorRef(
     if (rootElement !== null && resolution() !== null) {
       const { left, top, width, height } = resolution()!.getRect();
       const anchorHeight = anchorElementRef.offsetHeight; // use to position under anchor
-      containerDiv.style.top = `${top + window.pageYOffset + anchorHeight + 3}px`;
+      containerDiv.style.top = `${
+        top + window.pageYOffset + anchorHeight + 3
+      }px`;
       containerDiv.style.left = `${left + window.pageXOffset}px`;
       containerDiv.style.height = `${height}px`;
       containerDiv.style.width = `${width}px`;
@@ -500,7 +502,7 @@ export function useMenuAnchorRef(
         if (
           (top + menuHeight > window.innerHeight ||
             top + menuHeight > rootElementRect.bottom) &&
-          top - rootElementRect.top > menuHeight
+          top - rootElementRect.top > menuHeight + height
         ) {
           containerDiv.style.top = `${
             top - menuHeight + window.pageYOffset - height
