+++ lexical-solid/src/shared/useCharacterLimit.tsx
@@ -94,7 +94,6 @@ function findOffset(
   maxCharacters: number,
   strlen: (input: string) => number
 ): number {
-  // @ts-ignore This is due to be added in a later version of TS
   const Segmenter = Intl.Segmenter;
   let offsetUtf16 = 0;
   let offset = 0;
@@ -209,7 +208,7 @@ function $wrapOverflowedNodes(offset: number): void {
           $setSelection(previousSelection);
         }
 
-        mergePrevious(overflowNode);
+        $mergePrevious(overflowNode);
       }
     }
   }
@@ -217,7 +216,7 @@ function $wrapOverflowedNodes(offset: number): void {
 
 function $wrapNode(node: LexicalNode): OverflowNode {
   const overflowNode = $createOverflowNode();
-  node.insertBefore(overflowNode);
+  node.replace(overflowNode);
   overflowNode.append(node);
   return overflowNode;
 }
@@ -234,7 +233,7 @@ function $unwrapNode(node: OverflowNode): LexicalNode | null {
   return childrenLength > 0 ? children[childrenLength - 1] : null;
 }
 
-export function mergePrevious(overflowNode: OverflowNode): void {
+export function $mergePrevious(overflowNode: OverflowNode): void {
   const previousNode = overflowNode.getPreviousSibling();
 
   if (!$isOverflowNode(previousNode)) {
